% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_catch_by_age.R
\name{calc_catch_by_age}
\alias{calc_catch_by_age}
\title{Calculate Number of Fish Caught by Age}
\usage{
calc_catch_by_age(grain_key, grain_catch)
}
\arguments{
\item{grain_key}{A data frame with either wide-format columns (\code{Age-0}, \code{Age-1}, ...)
or long-format columns (\code{age}, \code{proportion}).}

\item{grain_catch}{A data frame output from \code{\link{calc_grain_catch}}, containing total number of fish
by \code{year}, \code{quarter}, and \code{Grain_class}.}
}
\value{
A data frame with number of fish caught by age, grain class, quarter, and year.
}
\description{
Estimates the number of fish caught by age class using a grain–age key
(either in wide or long format) and a table of total numbers caught per grain class.
}
