% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NV.R
\docType{data}
\name{data_NV}
\alias{data_NV}
\title{commercial landings data (simulated)}
\format{
A data frame with 4 columns:
\describe{
\item{FECHA_VENTA}{Date of sale (class Date)}
\item{TALLA}{Commercial grain size (factor or character)}
\item{TOTAL_EUROS}{Total value in euros (numeric)}
\item{TOTAL_KILOS}{Total landed weight in kilograms (numeric)}
}
}
\usage{
data(data_NV)
}
\description{
A simulated dataset representing commercial landings by grain size class.
}
\keyword{datasets}
