% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\docType{data}
\name{Database}
\alias{Database}
\title{Simulated  biological dataset}
\format{
A data frame with 3 columns:
\describe{
\item{operation_date}{Date of sampling (class Date)}
\item{edad}{Age of the individual (integer)}
\item{ptot_g}{Total weight in grams (numeric)}
}
}
\usage{
data(Database)
}
\description{
Example dataset of  individuals with age and total weight, used for building a grain–age key.
}
\keyword{datasets}
