% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grain_catch.R
\name{plot_grain_catch}
\alias{plot_grain_catch}
\title{Plot Number of Fish by Grain Class and Quarter}
\usage{
plot_grain_catch(grain_data)
}
\arguments{
\item{grain_data}{A data frame with columns:
\itemize{
\item \code{year}: Year of catch (numeric)
\item \code{quarter}: Quarter of catch (numeric)
\item \code{Grain_class}: Commercial grain class (character)
\item \code{num_total}: Estimated total number of fish caught
}}
}
\value{
A \code{ggplot2} bar plot (facetted by year-quarter)
}
\description{
Generates a bar plot of the estimated number of anchovies caught per grain class and quarter,
using the output from \code{\link{calc_grain_catch}}. Grain classes are automatically ordered
based on their labels (e.g., "31-50", "51-83").
}
