% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grain_key.R
\name{plot_grain_key}
\alias{plot_grain_key}
\title{Plot Grain Age Key}
\usage{
plot_grain_key(grain_key, plot_type = "percentage")
}
\arguments{
\item{grain_key}{A list returned by \code{build_grain_key()}, containing at least:
\code{grain_key} (proportion-at-age) and \code{weight_key} (mean weight-at-age).}

\item{plot_type}{Character string indicating which type of plot to return. Options:
\describe{
\item{"percentage"}{Plot proportion-at-age (default)}
\item{"mean_weight"}{Plot mean weight-at-age}
}}
}
\value{
A \code{ggplot2} object showing grain-age distribution or mean weight.
}
\description{
Visualizes either the proportion-at-age or the mean weight-at-age by grain class, year, and quarter,
using the output of \code{\link{build_grain_key}}.
}
\examples{
\dontrun{
results <- build_grain_key(data_bio)
plot_grain_key(results, plot_type = "percentage")
plot_grain_key(results, plot_type = "mean_weight")
}
}
